+++ pyodata/v2/service.py
@@ -267,10 +267,32 @@ class ODataHttpRequest:
         # pylint: disable=no-self-use
         return None
 
-    def get_headers(self):
-        """Get dict of HTTP headers"""
+    def get_default_headers(self):
+        """Get dict of Child specific HTTP headers"""
         # pylint: disable=no-self-use
-        return None
+        return dict()
+
+    def get_headers(self):
+        """Get dict of HTTP headers which is union of return value
+           of the method get_default_headers() and the headers
+           added via the method add_headers() where the latter
+           headers have priority - same keys get value of the latter.
+        """
+
+        headers = self.get_default_headers()
+        headers.update(self._headers)
+
+        return headers
+
+    def add_headers(self, value):
+        """Add the give dictionary of HTTP headers to
+           HTTP request sent by this ODataHttpRequest instance.
+        """
+
+        if not isinstance(value, dict):
+            raise TypeError("Headers must be of type 'dict' not {}".format(type(value)))
+
+        self._headers.update(value)
 
     def add_headers(self, value):
         if not isinstance(value, dict):
@@ -290,12 +312,7 @@ class ODataHttpRequest:
         # pylint: disable=assignment-from-none
         body = self.get_body()
 
-        headers = self._headers
-
-        # pylint: disable=assignment-from-none
-        extra_headers = self.get_headers()
-        if extra_headers is not None:
-            headers.update(extra_headers)
+        headers = self.get_headers()
 
         self._logger.debug('Send (execute) %s request to %s', self.get_method(), url)
         self._logger.debug('  query params: %s', self.get_query_params())
@@ -356,8 +373,8 @@ class EntityGetRequest(ODataHttpRequest):
     def get_path(self):
         return self._entity_set_proxy.last_segment + self._entity_key.to_key_string()
 
-    def get_headers(self):
-        return {'Accept': 'application/json', **self._headers}
+    def get_default_headers(self):
+        return {'Accept': 'application/json'}
 
     def get_query_params(self):
         qparams = super(EntityGetRequest, self).get_query_params()
@@ -453,8 +470,8 @@ class EntityCreateRequest(ODataHttpRequest):
     def get_body(self):
         return json.dumps(self._get_body())
 
-    def get_headers(self):
-        return {'Accept': 'application/json', 'Content-Type': 'application/json', 'X-Requested-With': 'X', **self._headers}
+    def get_default_headers(self):
+        return {'Accept': 'application/json', 'Content-Type': 'application/json', 'X-Requested-With': 'X'}
 
     @staticmethod
     def _build_values(entity_type, entity):
@@ -518,6 +535,8 @@ class EntityModifyRequest(ODataHttpRequest):
        Call execute() to send the update-request to the OData service
        and get the modified entity."""
 
+    ALLOWED_HTTP_METHODS = ['PATCH', 'PUT', 'MERGE']
+
     def __init__(self, url, connection, handler, entity_set, entity_key, method="PATCH"):
         super(EntityModifyRequest, self).__init__(url, connection, handler)
         self._logger = logging.getLogger(LOGGER_NAME)
@@ -525,10 +544,10 @@ class EntityModifyRequest(ODataHttpRequest):
         self._entity_type = entity_set.entity_type
         self._entity_key = entity_key
 
-        if method.upper() not in ["PATCH", "PUT"]:
-            raise ValueError("Method must be either PATCH or PUT")
-
-        self._method = method
+        self._method = method.upper()
+        if self._method not in EntityModifyRequest.ALLOWED_HTTP_METHODS:
+            raise ValueError('The value "{}" is not on the list of allowed Entity Update HTTP Methods: {}'
+                             .format(method, ', '.join(EntityModifyRequest.ALLOWED_HTTP_METHODS)))
 
         self._values = {}
 
@@ -551,8 +570,8 @@ class EntityModifyRequest(ODataHttpRequest):
             body[key] = val
         return json.dumps(body)
 
-    def get_headers(self):
-        return {'Accept': 'application/json', 'Content-Type': 'application/json', **self._headers}
+    def get_default_headers(self):
+        return {'Accept': 'application/json', 'Content-Type': 'application/json'}
 
     def set(self, **kwargs):
         """Set properties to be changed."""
@@ -644,7 +663,7 @@ class QueryRequest(ODataHttpRequest):
 
         return self._last_segment
 
-    def get_headers(self):
+    def get_default_headers(self):
         if self._count:
             return {}
 
@@ -708,10 +727,9 @@ class FunctionRequest(QueryRequest):
     def get_method(self):
         return self._function_import.http_method
 
-    def get_headers(self):
+    def get_default_headers(self):
         return {
-            'Accept': 'application/json',
-            **self._headers
+            'Accept': 'application/json'
         }
 
 
@@ -977,7 +995,6 @@ class GetEntitySetFilter:
 
     def __ge__(self, value):
         return GetEntitySetFilter.format_filter(self._proprty, 'ge', value)
-    
     def __gt__(self, value):
         return GetEntitySetFilter.format_filter(self._proprty, 'gt', value)
 
@@ -1165,7 +1182,7 @@ class EntitySetProxy:
         return EntityCreateRequest(self._service.url, self._service.connection, create_entity_handler, self._entity_set,
                                    self.last_segment)
 
-    def update_entity(self, key=None, method="PATCH", **kwargs):
+    def update_entity(self, key=None, method=None, **kwargs):
         """Updates an existing entity in the given entity-set."""
 
         def update_entity_handler(response):
@@ -1181,6 +1198,9 @@ class EntitySetProxy:
             entity_key = EntityKey(self._entity_set.entity_type, key, **kwargs)
 
         self._logger.info('Updating entity %s for key %s and args %s', self._entity_set.entity_type.name, key, kwargs)
+
+        if method is None:
+            method = self._service.config['http']['update_method']
 
         return EntityModifyRequest(self._service.url, self._service.connection, update_entity_handler, self._entity_set,
                                    entity_key, method=method)
@@ -1317,6 +1337,8 @@ class Service:
         self._entity_container = EntityContainer(self)
         self._function_container = FunctionContainer(self)
 
+        self._config = {'http': {'update_method': 'PATCH'}}
+
     @property
     def schema(self):
         """Parsed metadata"""
@@ -1346,6 +1368,12 @@ class Service:
         """Functions proxy"""
 
         return self._function_container
+
+    @property
+    def config(self):
+        """Service specific configuration"""
+
+        return self._config
 
     def http_get(self, path, connection=None):
         """HTTP GET response for the passed path in the service"""
@@ -1456,7 +1484,7 @@ class MultipartRequest(ODataHttpRequest):
         """Get boundary used for request parts"""
         return self.id
 
-    def get_headers(self):
+    def get_default_headers(self):
         # pylint: disable=no-self-use
         return {'Content-Type': 'multipart/mixed;boundary={}'.format(self.get_boundary()), **self._headers}