+++ CHANGELOG.md
@@ -7,7 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 
 ### Added
-- Specify PATCH or PUT method for EntityUpdateRequest - Barton Ip
+- Specify PATCH, PUT, or MERGE method for EntityUpdateRequest - Barton Ip
+- Add a Service wide configuration (e.g. http.update\_method) - Jakub Filak
 - <, <=, >, >= operators on GetEntitySetFilter - Barton Ip
 
 ### Fixed